requirements.txt
# pandas==2.1.0
# google-generativeai==0.3.1
# python-dotenv==1.0.0
# openai==1.3.0


# Core machine learning dependencies
torch>=2.0.0
transformers>=4.39.0
accelerate>=0.27.0

# For Qwen2.5 Omni model specifically
modelscope>=1.9.0
optimum>=1.18.0
peft>=0.7.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Image processing
Pillow>=10.0.0

# For API-based services
python-dotenv>=1.0.0
google-generativeai>=0.4.0
openai>=1.10.0

# File and environment management
safetensors>=0.4.0